requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
cryptography==41.0.7
python-dateutil==2.8.2
